Improve UI/UX and update polling intervals
- Update Flutter app polling interval from 10 minutes to 1 minute
- Fix BedCard layout to fit content within card height (h-64)
- Restore status badge colors to original light background style
- Fix progress bar visibility with inline backgroundColor styles
- Add patient info display to Flutter app header
- Change display from current weight to remaining percentage/volume
- Adjust spacing and padding to prevent content overflow

+++ Smart_IV_Pole-be/src/main/java/com/example/smartpole/entity/Prescription.java
@@ -28,7 +28,7 @@ public class Prescription {
     private Integer totalVolumeMl;
 
     @Column(name = "infusion_rate_ml_hr", nullable = false)
-    private Integer infusionRateMlHr; // 기존 필드 유지 (DB 호환성) - 실제로는 mL/min 저장됨
+    private Double infusionRateMlHr; // mL/min 단위로 저장 (Double로 소수점 유지)
 
     @Column(name = "gtt_factor", nullable = false)
     private Integer gttFactor; // 20 for macro, 60 for micro
